Move sidecar area preview inside blockedit
