+++ packages/block-library/src/blocks/sidecar-area/edit.js
@@ -1,16 +1,42 @@
-/**
- * WordPress dependencies
- */
+import classnames from 'classnames';
 
-import { Fragment } from '@wordpress/element';
+import {InnerBlocks} from '@wordpress/block-editor';
+import {dispatch, select} from '@wordpress/data';
 
-import SidecarAreaPreview from "./preview";
+const {getBlockRootClientId} = select( 'core/block-editor' );
+const {selectBlock} = dispatch( 'core/editor' );
 
-const SidecarAreaEdit = function (props) {
+const SidecarAreaEdit = function( props ) {
+
+  const {
+    attributes,
+    className,
+    clientId,
+    isSelected
+  } = props;
+
+  const parentClientId = getBlockRootClientId( clientId );
+
+  const {lastItemIsSticky} = attributes;
+
+  if ( isSelected ) {
+    selectBlock( parentClientId );
+  }
+
+  const classNames = classnames(
+    className,
+    {
+      'last-block-is-sticky': lastItemIsSticky === true
+    }
+  )
+
   return (
-    <Fragment>
-      <SidecarAreaPreview {...props} />
-    </Fragment>
+    <div className={classNames}>
+      <InnerBlocks
+        templateLock={false}
+        renderAppender={InnerBlocks.ButtonBlockAppender}
+      />
+    </div>
   )
 }
